% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makilabReg.R
\name{makilabReg}
\alias{makilabReg}
\title{Creates and exports a table of heirarchical LM models.}
\usage{
makilabReg(m1, ..., excel_export = FALSE, filename = NULL)
}
\arguments{
\item{m1}{An object output from lm() or a list of such objects.}

\item{...}{More models from lm().}

\item{excel_export}{A bool for whether not you want an excel export.}

\item{filename}{String of the filename to export to. Defaults to 'Data_YYYY-MM-DD.xlsx' where YYYY-MM-DD is today's date.}
}
\value{
The table of model summaries and (optional) an excel sheet added to the file.
}
\description{
Creates and exports a table of heirarchical LM models.
}
\examples{
data(iris)
library(makilab)
m1 <- lm(Sepal.Length~Petal.Length, data = iris)
m2 <- lm(Sepal.Length~Petal.Length+Petal.Width, data = iris)
m3 <- lm(Sepal.Length~Petal.Length+Petal.Width+Species, data = iris)
makilabReg(m1, m2, m3, excel_export = TRUE)
}
