% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makilabCor.R
\name{makilabCor}
\alias{makilabCor}
\title{Creates and exports a correlation matrix in a style consistent with the Maki Lab.}
\usage{
makilabCor(df, x, y, excel_export = FALSE, filename = NULL)
}
\arguments{
\item{df}{A data frame containing the variables you want to correlate.}

\item{x}{A list of column names for the X axis.}

\item{y}{A list of row names for the Y axis.}

\item{excel_export}{A bool for whether or not you want an excel export.}

\item{filename}{String of the filename to export to. Defaults to 'Data_YYYY-MM-DD.xlsx' where YYYY-MM-DD is today's date.}
}
\value{
The table of correlation values.
}
\description{
Creates and exports a correlation matrix in a style consistent with the Maki Lab.
}
\examples{
data(iris)
library(makilab)
makilabCor(df = iris,
           x = c("Sepal.Width", "Sepal.Length"),
           y = c("Petal.Width", "Petal.Length", "Species")) # No export
makilabCor(df = iris,
           x = c("Sepal.Width", "Sepal.Length"),
           y = c("Petal.Width", "Petal.Length", "Species"),
           excel_export = TRUE) # Exports to excel
}
