% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makilabReg.R
\name{makilabReg}
\alias{makilabReg}
\title{Creates and exports a table of LM models with multiple DVs and IVs.}
\usage{
makilabReg(m1, ..., excel_export = FALSE, filename = NULL, sheetname = NULL)
}
\arguments{
\item{m1}{An object output from lm() or a list of such objects.}

\item{...}{More models from lm().}

\item{excel_export}{A bool for whether not you want an excel export.}

\item{filename}{String of the filename to export to. Defaults to 'Data_YYYY-MM-DD.xlsx' where YYYY-MM-DD is today's date.}

\item{sheetname}{String to be used as the sheet name. Defaults to 'LMi' where i is the number of LMs in the file.}
}
\value{
The table of model summaries and (optional) an excel sheet added to the file.
}
\description{
Creates and exports a table of LM models with multiple DVs and IVs.
}
\examples{
data(iris)
library(makilab)
m1 <- lm(Petal.Length~Petal.Width+Species, data = iris)
m2 <- lm(Sepal.Length~Petal.Width+Sepal.Width, data = iris)
m3 <- lm(Petal.Width~Petal.Length+Sepal.Length+Species, data = iris)
makilabReg(m1, m2, m3, excel_export = TRUE, filename = "data.xlsx")
}
